Move LoopbackDispatcher to the demo extension as it is not a general implementation
